requirements.txt
streamlit>=1.24.0
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.23.0
langchain>=0.1.0
llama-index>=0.9.0
python-dotenv
